Drastically simplify implementation by removing customization
The customization of the class and namespace for the `AddServices` method added non-trivial complexity to the whole process.

Here we simplify it in two ways:

1 - In non-editor usage, the files are simply added as Compile, no need to process them
2 - In editor usage, we don't include them since we emit the [Obsolete] warning as needed.

This means we only need to do the additional file inclusion when building in the editor, which should minimize the impact of our generators on CI/CLI even more.